rearranging existing samples
