+++ example/lib/samples/disable_form_sample.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:reactive_forms_example/sample_screen.dart';
+
+class DisableFormSample extends StatefulWidget {
+  @override
+  _DisableFormSampleState createState() => _DisableFormSampleState();
+}
+
+class _DisableFormSampleState extends State<DisableFormSample> {
+  @override
+  Widget build(BuildContext context) {
+    return SampleScreen(
+      title: Text('Disable form sample'),
+    );
+  }
+}
